+++ packages/rax-icon/src/web/index.tsx
@@ -7,6 +7,7 @@ import Text from 'rax-text';
 import Image from 'rax-image';
 import wrapDefaultProperties from '../utils/wrapDefaultProperties';
 import { IconProps } from '../types';
+import { createIconSet as rawCreateIconSet } from '../createIconSet';
 
 const fontCache = new Map();
 const Icon: ForwardRefExoticComponent<IconProps> = forwardRef<HTMLSpanElement | HTMLImageElement, IconProps>(
@@ -55,5 +56,13 @@ const Icon: ForwardRefExoticComponent<IconProps> = forwardRef<HTMLSpanElement | 
     );
   }
 );
+
+export function createIconSet(
+  glyphMap = {},
+  fontFamily: string,
+  fontFile: string
+) {
+  return rawCreateIconSet(Icon, glyphMap, fontFamily, fontFile);
+}
 
 export default wrapDefaultProperties(Icon);