+++ packages/rax-icon/src/weex/index.tsx
@@ -6,6 +6,7 @@ import Text from 'rax-text';
 import Image from 'rax-image';
 import wrapDefaultProperties from '../utils/wrapDefaultProperties';
 import { IconProps } from '../types';
+import { createIconSet as rawCreateIconSet } from '../createIconSet';
 
 declare const __weex_require__: any;
 
@@ -53,5 +54,13 @@ const Icon = forwardRef<HTMLSpanElement | HTMLImageElement, IconProps>(
     );
   }
 );
+
+export function createIconSet(
+  glyphMap = {},
+  fontFamily: string,
+  fontFile: string
+) {
+  return rawCreateIconSet(Icon, glyphMap, fontFamily, fontFile);
+}
 
 export default wrapDefaultProperties(Icon);