+++ packages/rax-view/src/common/index.tsx
@@ -1,10 +1,8 @@
 import { createElement, useRef, forwardRef, ForwardRefExoticComponent } from 'rax';
 import cx from 'classnames/dedupe';
-import { isWeex } from 'universal-env';
 import wrapDefaultProperties from '../utils/wrapDefaultProperties';
 import { ViewProps } from '../types';
 import '../index.css';
-
 
 interface ViewRef {
   triggeredAppear: boolean;
@@ -27,7 +25,7 @@ const View: ForwardRefExoticComponent<ViewProps> = forwardRef(
         }
       };
     }
-    return <div {...rest} onAppear={handleAppear} ref={ref} className={cx( isWeex ? '' : 'rax-view-v2', className)} style={style} />;
+    return <div {...rest} onAppear={handleAppear} ref={ref} className={cx('rax-view-v2', className)} style={style} />;
   }
 );