+++ scripts/build.js
@@ -1,0 +1,8 @@
+const { readdirSync } = require('fs');
+const checkAndBuild = require('./check-deps-and-build');
+
+(() => {
+  const allPkgs = readdirSync('packages');
+
+  checkAndBuild(allPkgs);
+})();