+++ packages/rax-image/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
-## 2.4.1
+## 2.4.2
 
 - Support property `webp` in wechat-miniprogram
+
+## 2.4.1
+
+- Support placeholder in alibaba miniapp
 
 ## 2.4.0